Refactor wrapper and builder
Clean structure before introducing new builder implementations. Tie
builder to the wrapper to avoid cyclic imports.
