+++ pkg/gccbuilder/gcc-builder.go
@@ -1,16 +1,19 @@
-package builder
+package gccbuilder
 
 import (
 	"errors"
 	"fmt"
 	"log"
+	"os"
 	"path/filepath"
 	"regexp"
+	"runtime"
 	"sort"
 	"strings"
 
+	"github.com/QMSTR/qmstr/pkg/builder"
+	"github.com/QMSTR/qmstr/pkg/common"
 	pb "github.com/QMSTR/qmstr/pkg/service"
-	"github.com/QMSTR/qmstr/pkg/wrapper"
 	"github.com/spf13/pflag"
 )
 
@@ -94,61 +97,61 @@ type GccBuilder struct {
 	LinkLibs []string
 	LibPath  []string
 	Args     []string
-	GeneralBuilder
+	builder.GeneralBuilder
 }
 
 func NewGccBuilder(workDir string, logger *log.Logger, debug bool) *GccBuilder {
-	return &GccBuilder{Link, []string{}, []string{}, workDir, []string{}, []string{}, []string{}, GeneralBuilder{logger, debug}}
+	return &GccBuilder{Link, []string{}, []string{}, workDir, []string{}, []string{}, []string{}, builder.GeneralBuilder{logger, debug}}
 }
 
 func (g *GccBuilder) Analyze(commandline []string) (*pb.BuildMessage, error) {
-	if g.debug {
-		g.logger.Printf("Parsing commandline %v", commandline)
+	if g.Debug {
+		g.Logger.Printf("Parsing commandline %v", commandline)
 	}
 	g.parseCommandLine(commandline[1:])
 
 	switch g.Mode {
 	case Link:
-		g.logger.Printf("gcc linking")
+		g.Logger.Printf("gcc linking")
 		fileNodes := []*pb.FileNode{}
-		linkedTarget := NewFileNode(wrapper.BuildCleanPath(g.WorkDir, g.Output[0], false), linkedTrg)
+		linkedTarget := builder.NewFileNode(common.BuildCleanPath(g.WorkDir, g.Output[0], false), linkedTrg)
 		dependencies := []*pb.FileNode{}
 		for _, inFile := range g.Input {
 			inputFileNode := &pb.FileNode{}
 			ext := filepath.Ext(inFile)
 			if ext == ".o" {
-				inputFileNode = NewFileNode(wrapper.BuildCleanPath(g.WorkDir, inFile, false), obj)
+				inputFileNode = builder.NewFileNode(common.BuildCleanPath(g.WorkDir, inFile, false), obj)
 			} else if ext == ".c" {
-				inputFileNode = NewFileNode(wrapper.BuildCleanPath(g.WorkDir, inFile, false), src)
+				inputFileNode = builder.NewFileNode(common.BuildCleanPath(g.WorkDir, inFile, false), src)
 			} else {
-				inputFileNode = NewFileNode(wrapper.BuildCleanPath(g.WorkDir, inFile, false), linkedTrg)
+				inputFileNode = builder.NewFileNode(common.BuildCleanPath(g.WorkDir, inFile, false), linkedTrg)
 			}
 			dependencies = append(dependencies, inputFileNode)
 		}
-		actualLibs, err := wrapper.FindActualLibraries(g.LinkLibs, g.LibPath)
+		actualLibs, err := FindActualLibraries(g.LinkLibs, g.LibPath)
 		if err != nil {
-			g.logger.Fatalf("Failed to collect dependencies: %v", err)
+			g.Logger.Fatalf("Failed to collect dependencies: %v", err)
 		}
 		for _, actualLib := range actualLibs {
-			linkLib := NewFileNode(actualLib, linkedTrg)
+			linkLib := builder.NewFileNode(actualLib, linkedTrg)
 			dependencies = append(dependencies, linkLib)
 		}
 		linkedTarget.DerivedFrom = dependencies
 		fileNodes = append(fileNodes, linkedTarget)
 		return &pb.BuildMessage{FileNodes: fileNodes}, nil
 	case Assemble:
-		g.logger.Printf("gcc assembling - skipping link")
+		g.Logger.Printf("gcc assembling - skipping link")
 		fileNodes := []*pb.FileNode{}
-		if g.debug {
-			g.logger.Printf("This is our input %v", g.Input)
-			g.logger.Printf("This is our output %v", g.Output)
+		if g.Debug {
+			g.Logger.Printf("This is our input %v", g.Input)
+			g.Logger.Printf("This is our output %v", g.Output)
 		}
 		for idx, inFile := range g.Input {
-			if g.debug {
-				g.logger.Printf("This is the source file %s indexed %d", inFile, idx)
+			if g.Debug {
+				g.Logger.Printf("This is the source file %s indexed %d", inFile, idx)
 			}
-			sourceFile := NewFileNode(wrapper.BuildCleanPath(g.WorkDir, inFile, false), src)
-			targetFile := NewFileNode(wrapper.BuildCleanPath(g.WorkDir, g.Output[idx], false), obj)
+			sourceFile := builder.NewFileNode(common.BuildCleanPath(g.WorkDir, inFile, false), src)
+			targetFile := builder.NewFileNode(common.BuildCleanPath(g.WorkDir, g.Output[idx], false), obj)
 			targetFile.DerivedFrom = []*pb.FileNode{sourceFile}
 			fileNodes = append(fileNodes, targetFile)
 		}
@@ -162,20 +165,20 @@ func (g *GccBuilder) cleanCmdLine(args []string) {
 	clearIdxSet := map[int]struct{}{}
 	for idx, arg := range args {
 
-		if g.debug {
-			g.logger.Printf("%d - %s", idx, arg)
+		if g.Debug {
+			g.Logger.Printf("%d - %s", idx, arg)
 		}
 
 		// index string flags
 		if idx < len(args)-1 {
 			for key := range stringArgs {
-				if g.debug {
-					g.logger.Printf("Find %s string arg in %s with %s", key, fmt.Sprintf("%s %s ", arg, args[idx+1]), fmt.Sprintf("%s%s", key, stringArgsRE))
+				if g.Debug {
+					g.Logger.Printf("Find %s string arg in %s with %s", key, fmt.Sprintf("%s %s ", arg, args[idx+1]), fmt.Sprintf("%s%s", key, stringArgsRE))
 				}
 				re := regexp.MustCompile(fmt.Sprintf("%s%s", key, stringArgsRE))
 				if re.MatchString(fmt.Sprintf("%s %s ", arg, args[idx+1])) {
-					if g.debug {
-						g.logger.Printf("Found %v string arg", args[idx:idx+1])
+					if g.Debug {
+						g.Logger.Printf("Found %v string arg", args[idx:idx+1])
 					}
 					clearIdxSet[idx] = struct{}{}
 					clearIdxSet[idx+1] = struct{}{}
@@ -207,37 +210,37 @@ func (g *GccBuilder) cleanCmdLine(args []string) {
 	}
 	sort.Sort(sort.IntSlice(clear))
 
-	if g.debug {
-		g.logger.Printf("To be cleaned %v", clear)
+	if g.Debug {
+		g.Logger.Printf("To be cleaned %v", clear)
 	}
 	initialArgsSize := len(args)
 	for _, idx := range clear {
-		if g.debug {
-			g.logger.Printf("Clearing %d", idx)
+		if g.Debug {
+			g.Logger.Printf("Clearing %d", idx)
 		}
 		offset := initialArgsSize - len(args)
 		offsetIdx := idx - offset
-		if g.debug {
-			g.logger.Printf("Actually clearing %d", offsetIdx)
+		if g.Debug {
+			g.Logger.Printf("Actually clearing %d", offsetIdx)
 		}
 		if initialArgsSize-1 == idx {
-			if g.debug {
-				g.logger.Printf("Cut last arg")
+			if g.Debug {
+				g.Logger.Printf("Cut last arg")
 			}
 			args = args[:offsetIdx]
 		} else {
 			args = append(args[:offsetIdx], args[offsetIdx+1:]...)
 		}
-		if g.debug {
-			g.logger.Printf("new slice is %v", args)
+		if g.Debug {
+			g.Logger.Printf("new slice is %v", args)
 		}
 	}
 	g.Args = args
 }
 
 func (g *GccBuilder) parseCommandLine(args []string) {
-	if g.debug {
-		g.logger.Printf("Parsing arguments: %v", args)
+	if g.Debug {
+		g.Logger.Printf("Parsing arguments: %v", args)
 	}
 
 	// remove all flags we don't care about but that would break parsing
@@ -254,12 +257,12 @@ func (g *GccBuilder) parseCommandLine(args []string) {
 	gccFlags.StringSliceVarP(&g.LinkLibs, "linklib", "l", []string{}, "link libs")
 	gccFlags.StringSliceVarP(&g.LibPath, "linklibdir", "L", []string{}, "search dir for link libs")
 
-	if g.debug {
-		g.logger.Printf("Parsing cleaned commandline: %v", g.Args)
+	if g.Debug {
+		g.Logger.Printf("Parsing cleaned commandline: %v", g.Args)
 	}
 	err := gccFlags.Parse(g.Args)
 	if err != nil {
-		g.logger.Fatalf("Unrecoverable commandline parsing error: %s", err)
+		g.Logger.Fatalf("Unrecoverable commandline parsing error: %s", err)
 	}
 
 	g.Input = gccFlags.Args()
@@ -273,8 +276,8 @@ func (g *GccBuilder) parseCommandLine(args []string) {
 	if ok, err := gccFlags.GetBool("preprocess"); ok && err == nil {
 		g.Mode = Preproc
 	}
-	if g.debug {
-		g.logger.Printf("Mode set to: %v", g.Mode)
+	if g.Debug {
+		g.Logger.Printf("Mode set to: %v", g.Mode)
 	}
 
 	if output, err := gccFlags.GetString("output"); err == nil && output != undef {
@@ -296,4 +299,61 @@ func (g *GccBuilder) parseCommandLine(args []string) {
 			}
 		}
 	}
+}
+
+const (
+	libPathVar = "LIBRARY_PATH"
+)
+
+// FindActualLibraries discovers the actual libraries on the path
+func FindActualLibraries(libs []string, libpath []string) ([]string, error) {
+	actualLibPaths := []string{}
+	libpathvar, present := os.LookupEnv(libPathVar)
+	if present && libpathvar != "" {
+		libpath = append([]string{libpathvar}, libpath...)
+	}
+	var libprefix string
+	var libsuffix []string
+	var syslibpath []string
+	switch runtime.GOOS {
+	case "linux":
+		libprefix = "lib"
+		libsuffix = []string{".so"}
+		syslibpath = []string{"/lib", "/usr/lib", "/usr/local/lib", "/lib64"}
+	case "darwin":
+		libprefix = "lib"
+		libsuffix = []string{".dylib", ".so"}
+		syslibpath = []string{"/usr/lib", "/usr/local/lib"}
+	case "windows":
+		libprefix = ""
+		libsuffix = []string{".dll"}
+		syslibpath = []string{""}
+	}
+	for _, lib := range libs {
+		for _, dir := range append(libpath, syslibpath...) {
+			if dir == "" {
+				// Unix shell semantics: path element "" means "."
+				dir = "."
+			}
+			err := filepath.Walk(dir, func(path string, f os.FileInfo, err error) error {
+				for _, suffix := range libsuffix {
+					if f.Name() == fmt.Sprintf("%s%s%s", libprefix, lib, suffix) {
+						actualLibPaths = append(actualLibPaths, path)
+						return fmt.Errorf("Found %s", path)
+					}
+				}
+				return nil
+
+			})
+			if err != nil {
+				break
+			}
+		}
+	}
+
+	if len(actualLibPaths) == len(libs) {
+		return actualLibPaths, nil
+	}
+
+	return actualLibPaths, fmt.Errorf("Missing libraries from %v in %v", libs, actualLibPaths)
 }
